Improved item type styling and cascading in the List atoms.
